00_HACKATHON-SUBMISSIONS/0000_quiz-from-notes/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-dotenv==1.0.1
pdfplumber==0.11.4
requests==2.32.3